refactor: post, edit Component 분리
