+++ pages/freeboard/post.js
@@ -1,5 +1,6 @@
 import styles from '@/styles/Post.module.css';
-import { useEffect, useState } from 'react';
+import ArticleFormFields from '@/utils/ArticleFormFields';
+import { useState } from 'react';
 import axios from 'axios';
 import { useRouter } from 'next/router';
 import Button from '@/utils/Button';
@@ -7,7 +8,7 @@ import Button from '@/utils/Button';
 export default function Post() {
   const [titleValue, setTitleValue] = useState('');
   const [contentValue, setContentValue] = useState('');
-  const [submit, setSubmit] = useState(false);
+  const [canSubmit, setCanSubmit] = useState(false);
 
   const router = useRouter();
 
@@ -22,24 +23,11 @@ export default function Post() {
           userId: '3160c83b-8dcc-4ca2-9d51-717c5246d414',
         }
       );
-      console.log(res.data);
       router.push(`/article/${res.data.id}`);
     } catch (error) {
       console.error('Error posting data:', error);
     }
   }
-
-  const titleChange = (event) => {
-    const value = event.target.value;
-    setTitleValue(value);
-    setSubmit(value.trim() !== '' && contentValue.trim() !== '');
-  };
-
-  const contentChange = (event) => {
-    const value = event.target.value;
-    setContentValue(value);
-    setSubmit(value.trim() !== '' && titleValue.trim() !== '');
-  };
 
   function handleSubmit() {
     postArticle();
@@ -50,23 +38,14 @@ export default function Post() {
       <div className={styles.postLayout}>
         <div className={styles.header}>
           <span className={styles.title}>게시글 쓰기</span>
-          <Button disabled={!submit} onClick={handleSubmit} label={'등록'} />
+          <Button disabled={!canSubmit} onClick={handleSubmit} label={'등록'} />
         </div>
-
-        <div className={styles.name}>제목</div>
-        <input
-          placeholder='제목을 입력하세요'
-          value={titleValue}
-          onChange={titleChange}
-          id='titleValue'
-          className={styles.input}
-        />
-        <div className={styles.name}>내용</div>
-        <textarea
-          placeholder='내용을 입력하세요 '
-          value={contentValue}
-          onChange={contentChange}
-          className={styles.textarea}
+        <ArticleFormFields
+          titleValue={titleValue}
+          setTitleValue={setTitleValue}
+          contentValue={contentValue}
+          setContentValue={setContentValue}
+          setCanSubmit={setCanSubmit}
         />
       </div>
     </>
