+++ styles/Post.module.css
@@ -19,32 +19,3 @@
   line-height: 32px;
   color: #1f2937;
 }
-
-.name {
-  font-size: 18px;
-  font-weight: 700;
-  margin-bottom: 12px;
-}
-
-.input {
-  width: 100%;
-  border: none;
-  height: 56px;
-  background-color: #f3f4f6;
-  padding: 16px 24px;
-  border-radius: 12px;
-  box-sizing: border-box;
-  margin-bottom: 24px;
-}
-
-.textarea {
-  width: 100%;
-  border: none;
-  height: 282px;
-  background-color: #f3f4f6;
-  padding: 16px 24px;
-  box-sizing: border-box;
-  resize: none;
-  border-radius: 12px;
-  margin-bottom: 300px;
-}
